// src/_gen/api/auth/types.gen.ts
export type UserDataResponseDto = {
    id: string;
    email: string;
    name: string;
    roles: Array<'customer' | 'admin' | 'account-manager' | 'commercial' | 'ticket-reader'>;
};